Rename JsonCodec used by trino client
Having both io.trino.client.JsonCodec and io.airlift.json.JsonCodec
resulted in confusing mistakes.
