+++ client/trino-client/src/main/java/io/trino/client/auth/external/HttpTokenPoller.java
@@ -18,8 +18,8 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import dev.failsafe.Failsafe;
 import dev.failsafe.FailsafeException;
 import dev.failsafe.RetryPolicy;
-import io.trino.client.JsonCodec;
 import io.trino.client.JsonResponse;
+import io.trino.client.TrinoJsonCodec;
 import okhttp3.HttpUrl;
 import okhttp3.OkHttpClient;
 import okhttp3.Request;
@@ -36,8 +36,8 @@ import java.util.function.Supplier;
 import static com.google.common.base.MoreObjects.firstNonNull;
 import static com.google.common.net.HttpHeaders.USER_AGENT;
 import static io.trino.client.HttpStatusCodes.shouldRetry;
-import static io.trino.client.JsonCodec.jsonCodec;
 import static io.trino.client.JsonResponse.execute;
+import static io.trino.client.TrinoJsonCodec.jsonCodec;
 import static java.lang.String.format;
 import static java.net.HttpURLConnection.HTTP_INTERNAL_ERROR;
 import static java.net.HttpURLConnection.HTTP_OK;
@@ -47,7 +47,7 @@ import static java.util.Objects.requireNonNull;
 public class HttpTokenPoller
         implements TokenPoller
 {
-    private static final JsonCodec<TokenPollRepresentation> TOKEN_POLL_CODEC = jsonCodec(TokenPollRepresentation.class);
+    private static final TrinoJsonCodec<TokenPollRepresentation> TOKEN_POLL_CODEC = jsonCodec(TokenPollRepresentation.class);
     private static final String USER_AGENT_VALUE = "TrinoTokenPoller/" +
             firstNonNull(HttpTokenPoller.class.getPackage().getImplementationVersion(), "unknown");
