+++ client/trino-client/src/test/java/io/trino/client/TestQueryResults.java
@@ -18,13 +18,13 @@ import com.fasterxml.jackson.core.StreamReadConstraints;
 import com.google.common.base.Strings;
 import org.junit.jupiter.api.Test;
 
-import static io.trino.client.JsonCodec.jsonCodec;
+import static io.trino.client.TrinoJsonCodec.jsonCodec;
 import static java.lang.String.format;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class TestQueryResults
 {
-    private static final JsonCodec<QueryResults> QUERY_RESULTS_CODEC = jsonCodec(QueryResults.class);
+    private static final TrinoJsonCodec<QueryResults> QUERY_RESULTS_CODEC = jsonCodec(QueryResults.class);
 
     private static final String GOLDEN_VALUE = "{\n" +
             "  \"id\" : \"20160128_214710_00012_rk68b\",\n" +
