+++ client/trino-client/src/test/java/io/trino/client/TestRetry.java
@@ -34,8 +34,8 @@ import java.util.stream.Stream;
 
 import static com.google.common.net.HttpHeaders.CONTENT_TYPE;
 import static com.google.common.net.MediaType.JSON_UTF_8;
-import static io.trino.client.JsonCodec.jsonCodec;
 import static io.trino.client.StatementClientFactory.newStatementClient;
+import static io.trino.client.TrinoJsonCodec.jsonCodec;
 import static io.trino.spi.type.StandardTypes.INTEGER;
 import static io.trino.spi.type.StandardTypes.VARCHAR;
 import static java.lang.String.format;
@@ -48,7 +48,7 @@ import static org.junit.jupiter.api.TestInstance.Lifecycle.PER_METHOD;
 public class TestRetry
 {
     private MockWebServer server;
-    private static final JsonCodec<QueryResults> QUERY_RESULTS_CODEC = jsonCodec(QueryResults.class);
+    private static final TrinoJsonCodec<QueryResults> QUERY_RESULTS_CODEC = jsonCodec(QueryResults.class);
 
     @BeforeEach
     public void setup()
